+++ Core/libs/Helper.php
@@ -9,16 +9,9 @@ class Helper
 	public static function LoadConfig($file)
 	{
 		if(!file_exists($file)){
-
-			echo('//////////////////////////////////////');
-			Log::Error("config file does not exist".PHP_EOL);
-			echo('//////////////////////////////////////');
-			echo ($file.PHP_EOL);
-			echo "-------SOLUTION------".PHP_EOL.PHP_EOL;
-			Log::Info("Run: php builder cofig:init ".PHP_EOL.PHP_EOL);
-			echo "---------------------".PHP_EOL;
+			Logger::Error("Configuration file does not exist: $file");
+			Logger::Info("Please run: php vanilla config:init to generate the default configuration.");
 			die();
-
 		}
 		return Json::decode_file($file);
 	}
@@ -28,7 +21,7 @@ class Helper
 		Json::encode_file($file.'.conf', $data);
 		if(file_exists($file.'.conf')){
 			$name = explode('/', $file);
-			echo $name[count($name) - 1].".config file created successfully \n";
+			Logger::Success($name[count($name) - 1].".config file created successfully");
 		}
 	}
 	public static function isWin (){